Remove mongo-specific PermissionSet class
No good reason for it. Original thought was to prevent passing
permission sets from one implementation to another, but that serious
YAGNI BS.

TODO: make a builder, add unit tests.
